README.md
TLDR: Doesn't work yet, Makes me literally sick, caveat emptor.

Do not execute the code in the current commit, as looking at its animation literally made me sick. Like, haven't felt this sort of sick since playing Unreal Tournament 2004.

Other than that, I am proud to have rewritten the code in C++, so that it would run faster than javascript in the browser after being compiled to webassembly. MDN is a great resource on this topic, as are Google and Stackoverflow. I also relied on my Object Oriented Programming classes to create this.

The next steps are to create a web component, with support for vertex and edge elements nested inside of it, after reaching the milestone of returning a javascript object to accoplish the same things. I'm undecided as to whether there should be an html api apart from the fourd element. This will allow for a seamless integration of fourd into practically any website that chooses to run it, and any user agent allowing javascript once it is complete. The other step is to tweak the settings so that the simulation doesn't induce vomiting.

also, seizure warning!